+++ plugins/agentic-sdlc/src/agentic_sdlc/commands/__init__.py
@@ -5,10 +5,11 @@ from agentic_sdlc.commands.init import cmd_configure, cmd_init
 from agentic_sdlc.commands.release_notes import cmd_release_notes
 from agentic_sdlc.commands.resume import cmd_resume
 from agentic_sdlc.commands.run import cmd_run
-from agentic_sdlc.commands.shortcuts import cmd_analyze, cmd_input, cmd_oneshot
+from agentic_sdlc.commands.shortcuts import cmd_input
 from agentic_sdlc.commands.status import cmd_cancel, cmd_list, cmd_status
 from agentic_sdlc.commands.update import cmd_update
 from agentic_sdlc.commands.version import cmd_version
+from agentic_sdlc.commands.workflows import cmd_workflows
 
 __all__ = [
     "cmd_run",
@@ -19,10 +20,9 @@ __all__ = [
     "cmd_init",
     "cmd_configure",
     "cmd_config",
-    "cmd_oneshot",
-    "cmd_analyze",
     "cmd_input",
     "cmd_version",
     "cmd_release_notes",
     "cmd_update",
+    "cmd_workflows",
 ]